Changed how the gates interact with the gate input
The gates now have a reference to the gate inputs they are connected to and the gate inputs also have a reference to the gates they are connected to. There is still a game breaking bug that needs to be squashed and it is found in # TODOs: in the main.py file.